+++ packages/compiler/CHANGELOG.md
@@ -10,6 +10,16 @@
 #### Improvements
 
 - [#203](https://github.com/mesg-foundation/js-sdk/pull/203) Add support of any value for filters value (string, boolean or number)
+- [#](https://github.com/mesg-foundation/js-sdk/pull/) Add advance filter conditions
+A filter can now be defined as a list of conditions that contains a key, a predicate and a value. Previous definition of conditions with maps still works.
+eg: 
+```yaml
+  - type: filter
+    conditions:
+      - key: data.foo[0].x
+        predicate: EQ
+        value: 2
+```
 
 #### Bug fixes